+++ src/mode/visual_mode.rs
@@ -462,6 +462,7 @@ mod tests {
             fields: Default::default(),
             levels: Default::default(),
             multiline: false,
+            continuation: None,
         };
         tab.display.format = Some(std::sync::Arc::new(
             crate::parser::CustomParser::from_config(&cfg).unwrap(),
@@ -497,6 +498,7 @@ mod tests {
                 .collect(),
             levels: Default::default(),
             multiline: false,
+            continuation: None,
         };
         tab.display.format = Some(std::sync::Arc::new(
             crate::parser::CustomParser::from_config(&cfg).unwrap(),
@@ -698,6 +700,7 @@ mod tests {
             fields: Default::default(),
             levels: Default::default(),
             multiline: false,
+            continuation: None,
         };
         tab.display.format = Some(std::sync::Arc::new(
             crate::parser::CustomParser::from_config(&cfg).unwrap(),
@@ -733,6 +736,7 @@ mod tests {
                 .collect(),
             levels: Default::default(),
             multiline: false,
+            continuation: None,
         };
         tab.display.format = Some(std::sync::Arc::new(
             crate::parser::CustomParser::from_config(&cfg).unwrap(),